ANGULAR PROJECTS/materialinv-django/requirements.txt
django
django-crispy-forms
crispy-bootstrap4
